edit script to be architecture agnostic
all script which previously had hardcoded arm64/aarch64 now detect the host architecture

for programs/applications that support it, the megascript should now be safe and functional to use on aarch64, armhf, i386, and x86_64

+++ scripts/sdl2_install_helper.sh
@@ -14,42 +14,48 @@ if [ $(version $sdlv) -ge $(version "2.0.14.0") ]; then
     echo "Already Installed Newest SDL2 Version"
     sleep 1
 else
-    cd /tmp
-    wget https://github.com/$repository_username/L4T-Megascript/raw/$repository_branch/assets/SDL2/libsdl2-2.0-0_2.0.14%2B5_arm64.deb
-    wget https://github.com/$repository_username/L4T-Megascript/raw/$repository_branch/assets/SDL2/libsdl2-dbg_2.0.14%2B5_arm64.deb
-    wget https://github.com/$repository_username/L4T-Megascript/raw/$repository_branch/assets/SDL2/libsdl2-dev_2.0.14%2B5_arm64.deb
+    get_system
+    if [[ "$architecture" == "aarch64" ]]; then
+        cd /tmp
+        wget https://github.com/$repository_username/L4T-Megascript/raw/$repository_branch/assets/SDL2/libsdl2-2.0-0_2.0.14%2B5_arm64.deb
+        wget https://github.com/$repository_username/L4T-Megascript/raw/$repository_branch/assets/SDL2/libsdl2-dbg_2.0.14%2B5_arm64.deb
+        wget https://github.com/$repository_username/L4T-Megascript/raw/$repository_branch/assets/SDL2/libsdl2-dev_2.0.14%2B5_arm64.deb
 
-    sudo dpkg -i libsdl2-dev_2.0.14+5_arm64.deb libsdl2-dbg_2.0.14+5_arm64.deb libsdl2-2.0-0_2.0.14+5_arm64.deb
-    sudo apt install -fy
-    sudo rm -rf libsdl2-dev_2.0.14+5_arm64.deb libsdl2-dbg_2.0.14+5_arm64.deb libsdl2-2.0-0_2.0.14+5_arm64.deb
+        sudo dpkg -i libsdl2-dev_2.0.14+5_arm64.deb libsdl2-dbg_2.0.14+5_arm64.deb libsdl2-2.0-0_2.0.14+5_arm64.deb
+        sudo apt install -fy
+        sudo rm -rf libsdl2-dev_2.0.14+5_arm64.deb libsdl2-dbg_2.0.14+5_arm64.deb libsdl2-2.0-0_2.0.14+5_arm64.deb
 
-    # keep this old code for compiling SDL2 if its ever needed
+        # keep this old code for compiling SDL2 if its ever needed
 
-    # sudo apt-get --assume-yes install git
-    # cd /tmp
-    # mkdir -p temp_install_sdl2
-    # cd temp_install_sdl2
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/retropie_packages.sh' > retropie_packages.sh
-    # chmod -R 777 retropie_packages.sh
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/retropie_setup.sh' > retropie_setup.sh
-    # mkdir -p scriptmodules
-    # cd ./scriptmodules
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/system.sh' > system.sh
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/helpers.sh' > helpers.sh
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/packages.sh' > packages.sh
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/inifuncs.sh' > inifuncs.sh
-    # mkdir -p supplementary
-    # cd ./supplementary
-    # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/supplementary/sdl2.sh' > sdl2.sh
-    # cd /tmp/temp_install_sdl2
+        # sudo apt-get --assume-yes install git
+        # cd /tmp
+        # mkdir -p temp_install_sdl2
+        # cd temp_install_sdl2
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/retropie_packages.sh' > retropie_packages.sh
+        # chmod -R 777 retropie_packages.sh
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/retropie_setup.sh' > retropie_setup.sh
+        # mkdir -p scriptmodules
+        # cd ./scriptmodules
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/system.sh' > system.sh
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/helpers.sh' > helpers.sh
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/packages.sh' > packages.sh
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/inifuncs.sh' > inifuncs.sh
+        # mkdir -p supplementary
+        # cd ./supplementary
+        # curl -s 'https://raw.githubusercontent.com/theofficialgman/RetroPie-Setup/master/scriptmodules/supplementary/sdl2.sh' > sdl2.sh
+        # cd /tmp/temp_install_sdl2
 
-    # #auto install sdl2 and then remove unneeded files
-    # sudo ./retropie_packages.sh sdl2
-    # cd /tmp
-    # sudo rm -rf temp_install_sdl2
+        # #auto install sdl2 and then remove unneeded files
+        # sudo ./retropie_packages.sh sdl2
+        # cd /tmp
+        # sudo rm -rf temp_install_sdl2
 
-    cd
-    echo ""
-    echo "Successfully Installed Newest SDL2 Version"
-    sleep 3
+        cd
+        echo ""
+        echo "Successfully Installed Newest SDL2 Version"
+        sleep 3
+    else
+        echo "Sorry we don't host binaries on the megascript for non-aarch64 architectures"
+        echo "Skipping updated SDL2 install"
+    fi
 fi
